requirements.txt
discord.py
python-dotenv
tabulate
emoji
